kafka-druid-ingest/requirements.txt
kafka-python
requests<2.31
urllib3<2